extracted UseProductImageUrl Hook Hook
