+++ src/app/[locale]/viewer/_components/AASOverviewCard.tsx
@@ -8,7 +8,7 @@ import {
     Skeleton,
     Typography,
 } from '@mui/material';
-import React, { useState } from 'react';
+import React from 'react';
 import { DataRow } from 'components/basics/DataRow';
 import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
 import { AssetAdministrationShell, SpecificAssetId } from '@aas-core-works/aas-core3.0-typescript/types';
@@ -17,13 +17,11 @@ import { AssetIcon } from 'components/custom-icons/AssetIcon';
 import { ShellIcon } from 'components/custom-icons/ShellIcon';
 import { isValidUrl } from 'lib/util/UrlUtil';
 import { encodeBase64 } from 'lib/util/Base64Util';
-import { useAsyncEffect } from 'lib/hooks/UseAsyncEffect';
-import { useRouter } from 'next/navigation';
 import { useAasState } from 'components/contexts/CurrentAasContext';
 import { ImageWithFallback } from 'components/basics/StyledImageWithFallBack';
-import { getThumbnailFromShell } from 'lib/services/repository-access/repositorySearchActions';
-import { mapFileDtoToBlob } from 'lib/util/apiResponseWrapper/apiResponseWrapper';
 import { useTranslations } from 'next-intl';
+import { useRouter } from 'next/navigation';
+import { useProductImageUrl } from 'lib/hooks/UseProductImageUrl';
 
 type AASOverviewCardProps = {
     readonly aas: AssetAdministrationShell | null;
@@ -58,36 +56,9 @@ export function AASOverviewCard(props: AASOverviewCardProps) {
     const isAccordion = props.isAccordion;
     const specificAssetIds = props.aas?.assetInformation?.specificAssetIds as SpecificAssetId[];
     const navigate = useRouter();
-    const [productImageUrl, setProductImageUrl] = useState<string>('');
     const [, setAasState] = useAasState();
     const t = useTranslations('pages.aasViewer');
-
-    async function createAndSetUrlForImageFile() {
-        if (!props.aas) return;
-
-        if (!props.repositoryURL) {
-            setProductImageUrl('');
-            return;
-        }
-
-        const response = await getThumbnailFromShell(props.aas.id, props.repositoryURL);
-        if (!response.isSuccess) {
-            console.error('Image not found');
-            return;
-        }
-        const blob = mapFileDtoToBlob(response.result);
-        setProductImageUrl(URL.createObjectURL(blob));
-    }
-
-    useAsyncEffect(async () => {
-        if (!props.productImage) return;
-
-        if (!isValidUrl(props.productImage!)) {
-            await createAndSetUrlForImageFile();
-        } else {
-            setProductImageUrl(props.productImage);
-        }
-    }, [props.productImage]);
+    const productImageUrl = useProductImageUrl(props.aas, props.repositoryURL);
 
     const infoBoxStyle = {
         display: 'flex',
